// front-end/src/modules/admin/AdminJournals/journalList.types.ts
export interface ICurrency {
  company_id(arg0: string, company_id: any): unknown;
  bank_account_id(arg0: string, bank_account_id: any): unknown;
  company_name(arg0: string, company_name: any): unknown;
  currency_name: string;
  id: string;
  short_code: string;
  status: string;
  symbol: string;
}